File structure updated - input class file and additional features file. Abstract print now inherits from life and death. life and death generic checker now checks if the set is typed tuple or int
